# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Moskollektor Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/facilities": {
      "get": {
        "tags": [
          "facilities"
        ],
        "summary": "Get Facilities",
        "description": "List facilities within the caller's scope.\n\nArgs:\n    query: Optional case-insensitive substring match on display_name.\n    current_state: Accepted for contract completeness (see facility_query).\n    risk_level: Accepted for contract completeness (see facility_query).\n    district: Optional district_id filter.\n    bbox: Optional \"min_lon,min_lat,max_lon,max_lat\" filter.\n    cursor: Opaque pagination cursor.\n    limit: Page size, 1-200.\n    user: The authenticated caller, injected by get_current_user.\n\nReturns:\n    A FacilityListEnvelope containing only facilities the caller may see.",
        "operationId": "get_facilities_api_v1_facilities_get",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Query"
            }
          },
          {
            "name": "current_state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Current State"
            }
          },
          {
            "name": "risk_level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Risk Level"
            }
          },
          {
            "name": "district",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "District"
            }
          },
          {
            "name": "bbox",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bbox"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FacilityListEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/facilities/{facility_id}": {
      "get": {
        "tags": [
          "facilities"
        ],
        "summary": "Get Facility",
        "description": "Fetch one facility's detail, enforcing the caller's scope.\n\nArgs:\n    facility_id: The requested facility id.\n    user: The authenticated caller, injected by get_current_user.\n\nReturns:\n    The facility's full detail (same shape as a list item).\n\nRaises:\n    ApiError: 403 if the facility exists but is out of scope; 404 if\n        it does not exist.",
        "operationId": "get_facility_api_v1_facilities__facility_id__get",
        "parameters": [
          {
            "name": "facility_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Facility Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FacilityOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Assets": {
        "properties": {
          "collector_count": {
            "type": "integer",
            "title": "Collector Count"
          },
          "sensor_count": {
            "type": "integer",
            "title": "Sensor Count"
          },
          "offline_sensor_count": {
            "type": "integer",
            "title": "Offline Sensor Count"
          }
        },
        "type": "object",
        "required": [
          "collector_count",
          "sensor_count",
          "offline_sensor_count"
        ],
        "title": "Assets",
        "description": "Counts of the facility's owned equipment/sensors."
      },
...
      "DataHealth": {
        "properties": {
          "freshness": {
            "type": "string",
            "title": "Freshness"
          },
          "last_event_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Event At"
          },
          "coverage": {
            "type": "number",
            "title": "Coverage"
          }
        },
        "type": "object",
        "required": [
          "freshness",
          "last_event_at",
          "coverage"
        ],
        "title": "DataHealth",
        "description": "Freshness/coverage of the data backing this facility's other fields."
      },
...
      "FacilityListEnvelope": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/FacilityOut"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "$ref": "#/components/schemas/FacilityListMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "FacilityListEnvelope",
        "description": "The standard {data, meta} envelope for GET /api/v1/facilities."
      },
      "FacilityListMeta": {
        "properties": {
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          }
        },
        "type": "object",
        "required": [
          "next_cursor",
          "total",
          "generated_at"
        ],
        "title": "FacilityListMeta",
        "description": "Cursor-pagination metadata for the facility list envelope."
      },
      "FacilityOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "facility_type": {
            "type": "string",
            "title": "Facility Type"
          },
          "location": {
            "$ref": "#/components/schemas/Location"
          },
          "current_state": {
            "type": "string",
            "title": "Current State"
          },
          "forecast": {
            "$ref": "#/components/schemas/Forecast"
          },
          "incidents": {
            "$ref": "#/components/schemas/Incidents"
          },
          "assets": {
            "$ref": "#/components/schemas/Assets"
          },
          "data_health": {
            "$ref": "#/components/schemas/DataHealth"
          },
          "priority_score": {
            "type": "number",
            "title": "Priority Score"
          },
          "source_health": {
            "items": {
              "$ref": "#/components/schemas/SourceHealthEntry"
            },
            "type": "array",
            "title": "Source Health"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "display_name",
          "facility_type",
          "location",
          "current_state",
          "forecast",
          "incidents",
          "assets",
          "data_health",
          "priority_score",
          "source_health",
          "updated_at"
        ],
        "title": "FacilityOut",
        "description": "The complete facility representation, shared by list and detail."
      },
      "Forecast": {
        "properties": {
          "risk_level": {
            "type": "string",
            "title": "Risk Level"
          },
          "max_probability": {
            "type": "number",
            "title": "Max Probability"
          },
          "active_count": {
            "type": "integer",
            "title": "Active Count"
          },
          "earliest_window_start": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Earliest Window Start"
          }
        },
        "type": "object",
        "required": [
          "risk_level",
          "max_probability",
          "active_count",
          "earliest_window_start"
        ],
        "title": "Forecast",
        "description": "Forecast summary — independent from current_state and incidents."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Incidents": {
        "properties": {
          "open_count": {
            "type": "integer",
            "title": "Open Count"
          },
          "critical_count": {
            "type": "integer",
            "title": "Critical Count"
          }
        },
        "type": "object",
        "required": [
          "open_count",
          "critical_count"
        ],
        "title": "Incidents",
        "description": "Confirmed-incident counters — independent from forecast/current_state."
      },
      "Location": {
        "properties": {
          "type": {
            "type": "string",
            "const": "Point",
            "title": "Type",
            "default": "Point"
          },
          "coordinates": {
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "type": "array",
            "maxItems": 2,
            "minItems": 2,
            "title": "Coordinates"
          },
          "is_demo": {
            "type": "boolean",
            "title": "Is Demo",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "coordinates"
        ],
        "title": "Location",
        "description": "A GeoJSON Point with an explicit demo-geometry flag.\n\nReal coordinates will never be provided (see\n`QA_ORGANIZERS_CLARIFICATIONS.md` section 12); this point is\ndeterministic (stable per facility_id) but not geographically real."
      },
...
      "SourceHealthEntry": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "last_success_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Success At"
          },
          "delay_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delay Seconds"
          }
        },
        "type": "object",
        "required": [
          "source",
          "display_name",
          "status",
          "last_success_at",
          "delay_seconds"
        ],
        "title": "SourceHealthEntry",
        "description": "Status of one upstream data source (see ticket 06 emulation module)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}